modified:   ai-stack/ai-services/compose-up.sh 	modified:   scripts/create_networks.sh 	modified:   scripts/create_secrets.sh 	modified:   scripts/custom-launch-docker.sh 	modified:   scripts/install_drivers.sh 	modified:   scripts/install_toolhive.sh

+++ scripts/create_secrets.sh
@@ -29,7 +29,9 @@ function CREATE_SECRETS(){
     if [[ ! -f "${WD}/ai-stack/DATA/anythingllm/anythingllm_storage/.env" ]]; then
         mkdir -p "${WD}/ai-stack/DATA/anythingllm/anythingllm_storage"
         cd "${WD}/ai-stack/DATA/anythingllm/anythingllm_storage" || exit 1
-        sudo wget -c "https://raw.githubusercontent.com/Mintplex-Labs/anything-llm/refs/heads/master/server/.env.example" -O ".env"
+        if [[ ! -f ".env" ]]; then
+            sudo wget -c "https://raw.githubusercontent.com/Mintplex-Labs/anything-llm/refs/heads/master/server/.env.example" -O ".env"
+        fi
     fi
 }
 CREATE_SECRETS
